+++ scripts/fetch-iana.js
@@ -19,7 +19,6 @@ var toArray = require('stream-to-array')
 var writedb = require('./lib/write-db')
 
 var extensionsQuotedRegExp = /^\s*(?:\d\.\s+)?File extension(?:\(s\)|s|)\s?:(?:[^'"\r\n]+)(?:"\.?([0-9a-z_-]+)"|'\.?([0-9a-z_-]+)')/im
-var intendedUsageRegExp = /^\s*(?:(?:\d{1,2}\.|o)\s+)?Intended\s+Usage\s*:\s*(\S+)/im
 var leadingSpacesRegExp = /^\s+/
 var listColonRegExp = /:(?:\s|$)/m
 var nameWithNotesRegExp = /^(\S+)(?: - (.*)$| \((.*)\)$|)/
@@ -33,6 +32,7 @@ var urlReferenceRegExp = /\[(https?:\/\/[^\]]+)]/gi
 
 var CHARSET_DEFAULT_REGEXP = /(?:\bcharset\b[^.]*(?:\.\s+default\s+(?:value\s+)?is|\bdefault[^.]*(?:of|is)|\bmust\s+have\s+the\s+value|\bvalue\s+must\s+be)\s+|\bcharset\s*\(?defaults\s+to\s+|\bdefault\b[^.]*?\bchar(?:set|act[eo]r\s+set)\b[^.]*?(?:of|is)\s+|\bcharset\s+(?:must|is)\s+always\s+(?:be\s+)?)["']?([a-z0-9]+-[a-z0-9-]+)/im
 var EXTENSIONS_REGEXP = /(?:^\s*(?:\d\.\s+)?|\s+[23]\.\s+)[Ff]ile [Ee]xtension(?:\(s\)|s|)\s?:\s+(?:\*\.|\.|)([0-9a-z_-]+|[0-9A-Z_-]+)(?:\s+or\s+(?:\*\.|\.|)([0-9a-z_-]+|[0-9A-Z_-]+)\s*)?(?:\s*[34]\.\s+|\s+[A-Z(]|\s*$)/m
+var INTENDED_USAGE_REGEXP = /^\s*(?:(?:\d{1,2}\.|o)\s+)?Intended\s+Usage\s*:\s*([0-9a-z]+)/im
 var MIME_SUBTYPE_LINE_REGEXP = /^[^:\s-]*\s*(?:MIME )?(?:[Mm]edia )?(?:[Ss]ub ?type|SUB ?TYPE)(?: (?:[Nn]ame|NAME))?\s*:\s+(?:[A-Za-z]+ [Tt]ree\s+(?:- ?)?|(?:[a-z]+ )+- )?([0-9A-Za-z][0-9A-Za-z_.+-]*)(?:\s|$)/m
 var MIME_TYPE_HAS_CHARSET_PARAMETER_REGEXP = /parameters\s*:[^.]*\bcharset\b/im
 
@@ -76,6 +76,11 @@ co(function * () {
 
     if (mime in json) {
       throw new Error('duplicate entry for ' + mime)
+    }
+
+    // skip obsoleted mimes
+    if (result.usage === 'obsolete') {
+      return
     }
 
     json[mime] = {
@@ -145,10 +150,12 @@ function addTemplateData (data, options) {
       // use extracted charset
       data.charset = extractTemplateCharset(body)
 
+      // use extracted usage
+      data.usage = extractIntendedUsage(body)
+
       // use extracted extensions
-      var useExt = opts.extensions &&
-        (opts.extensions === true || opts.extensions.test(data.mime))
-      if (useExt && extractIntendedUsage(body) === 'common') {
+      if (data.usage === 'common' && opts.extensions &&
+        (opts.extensions === true || opts.extensions.test(data.mime))) {
         data.extensions = extractTemplateExtensions(body)
       }
     }
@@ -158,7 +165,7 @@ function addTemplateData (data, options) {
 }
 
 function extractIntendedUsage (body) {
-  var match = intendedUsageRegExp.exec(body)
+  var match = INTENDED_USAGE_REGEXP.exec(body)
 
   return match
     ? match[1].toLocaleLowerCase()